Move use*Modal to shared/ui, fix after rebasing
