+++ src/features/cart/addToCart/ui/RemoveIcon/RemoveIcon.tsx
@@ -1,8 +1,7 @@
 import { useCallback } from 'react'
 import type { ProductId } from '@/entities/product'
-import { useConfirmModal } from '@/shared/lib'
 import { useAppDispatch } from '@/shared/redux'
-import { Icon } from '@/shared/ui'
+import { Icon, useConfirmModal } from '@/shared/ui'
 import { removeCartItemThunk } from '../../model/actions'
 
 type Props = {
