+++ src/shared/lib/index.ts
@@ -6,5 +6,4 @@ export {
 } from './jwt'
 export { wait } from './wait'
 export { debounce } from './debounce'
-export { useConfirmModal } from './useConfirmModal'
 export { getQueryParams, useTypedParams } from './router'
