+++ src/shared/ui/index.ts
@@ -4,3 +4,5 @@ export { Button } from './Button/Button'
 export { Input } from './Input/Input'
 export { Modal } from './Modal/Modal'
 export { PageHeader } from './PageHeader/PageHeader'
+export { useConfirmModal } from './useConfirmModal'
+export { useAlertModal } from './useAlertModal'
